+++ packages/native/src/components/Column/index.ts
@@ -1,11 +1,14 @@
 import { ComponentProps } from 'react'
 import styled from 'styled-components/native'
 import { getHorizontalAlignment, getVerticalAlignment } from '@shallot-ui/core'
+import { INTERNAL_shouldForwardProp } from '@shallot-ui/theme'
 
 import { Box } from '../Box'
 import { ColumnStyleProps } from './style/types'
 
-export const Column = styled(Box)<ColumnStyleProps>`
+export const Column = styled(Box).withConfig<ColumnStyleProps>({
+  shouldForwardProp: INTERNAL_shouldForwardProp,
+})`
   flex-direction: column;
   ${getVerticalAlignment('column')}
   ${getHorizontalAlignment('column')}