+++ packages/native/src/components/Input/base.ts
@@ -1,3 +1,4 @@
+import { INTERNAL_shouldForwardProp } from '@shallot-ui/theme'
 import { ComponentProps } from 'react'
 import { TextInput } from 'react-native'
 import styled from 'styled-components'
@@ -5,11 +6,13 @@ import styled from 'styled-components'
 import { getInputStyle } from './getStyle'
 import { InputStyleProps } from './types'
 
-export const Input = styled(TextInput).attrs<InputStyleProps>(
-  ({ placeholderColor, theme }) => ({
+export const Input = styled(TextInput)
+  .attrs<InputStyleProps>(({ placeholderColor, theme }) => ({
     placeholderTextColor: placeholderColor && theme.colors[placeholderColor],
-  }),
-)`
+  }))
+  .withConfig({
+    shouldForwardProp: INTERNAL_shouldForwardProp,
+  })`
   display: flex;
   border-style: solid;
   outline: none;