+++ packages/native/src/components/Row/index.ts
@@ -4,8 +4,11 @@ import { getHorizontalAlignment, getVerticalAlignment } from '@shallot-ui/core'
 
 import { Box } from '../Box'
 import { RowStyleProps } from './style/types'
+import { INTERNAL_shouldForwardProp } from '@shallot-ui/theme'
 
-export const Row = styled(Box)<RowStyleProps>`
+export const Row = styled(Box).withConfig<RowStyleProps>({
+  shouldForwardProp: INTERNAL_shouldForwardProp,
+})`
   flex-direction: row;
   ${getVerticalAlignment('row')}
   ${getHorizontalAlignment('row')}