+++ packages/native/src/components/Space/index.ts
@@ -1,6 +1,9 @@
+import { INTERNAL_shouldForwardProp } from '@shallot-ui/theme'
 import styled from 'styled-components/native'
-import { Box } from '../Box'
+import { Box, BoxStyleProps } from '../Box'
 
-export const Space = styled(Box)`
+export const Space = styled(Box).withConfig<BoxStyleProps>({
+  shouldForwardProp: INTERNAL_shouldForwardProp,
+})`
   flex-grow: 1;
 `