+++ packages/web/src/components/Space/index.ts
@@ -1,7 +1,10 @@
 import styled from 'styled-components'
+import { INTERNAL_shouldForwardProp } from '@shallot-ui/theme'
 
-import { Box } from '../primitives'
+import { Box, BoxStyleProps } from '../primitives'
 
-export const Space = styled(Box)`
+export const Space = styled(Box).withConfig<BoxStyleProps>({
+  shouldForwardProp: INTERNAL_shouldForwardProp,
+})`
   flex: 1;
 `