+++ packages/web/src/components/primitives/Column/index.ts
@@ -1,3 +1,4 @@
+import { INTERNAL_shouldForwardProp } from '@shallot-ui/theme'
 import { ComponentProps } from 'react'
 import styled from 'styled-components'
 
@@ -5,7 +6,9 @@ import { Box } from '../Box'
 import { getColumnStyle } from './style/get'
 import { ColumnStyleProps } from './style/types'
 
-export const Column = styled(Box)<ColumnStyleProps>`
+export const Column = styled(Box).withConfig<ColumnStyleProps>({
+  shouldForwardProp: INTERNAL_shouldForwardProp,
+})`
   ${getColumnStyle}
 `