+++ packages/web/src/components/primitives/Row/index.ts
@@ -1,11 +1,14 @@
 import { ComponentProps } from 'react'
 import styled from 'styled-components'
+import { INTERNAL_shouldForwardProp } from '@shallot-ui/theme'
 
 import { Box } from '../Box'
 import { getRowStyle } from './style/get'
 import { RowStyleProps } from './style/types'
 
-export const Row = styled(Box)<RowStyleProps>`
+export const Row = styled(Box).withConfig<RowStyleProps>({
+  shouldForwardProp: INTERNAL_shouldForwardProp,
+})`
   ${getRowStyle}
 `